requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
python-multipart>=0.0.9
jinja2>=3.1.0
openpyxl>=3.1.0
gunicorn>=22.0.0
psycopg[binary]>=3.2.0
pytest>=8.0.0
